requirements.txt
pyserial
# esptool
pyqt6
requests